make BPM fractional, fixes #1

+++ helper.go
@@ -7,7 +7,7 @@ import (
 
 // tempoBasedOnMIDIClocks returns a tempo calculated by 4 timestamps of midi clock events in a row
 // a is the first and d is the last timestamp
-func tempoBasedOnMIDIClocks(a, b, c, d *time.Time) uint32 {
+func tempoBasedOnMIDIClocks(a, b, c, d *time.Time) float64 {
 	// the simplest way to do this is to build the difference in time to the last midi clock
 	// 12 clocks = 8th, 6 = 16th, 3 = 32th
 	// lets say we always takt the last 3 midi clocks their timespan must be a 32th note
@@ -21,5 +21,5 @@ func tempoBasedOnMIDIClocks(a, b, c, d *time.Time) uint32 {
 	bebefore := b.Sub(*a)
 	one32th := (last + before + bebefore)
 	tempo := (time.Second * 60 / (one32th * 8)).Nanoseconds()
-	return uint32(tempo)
+	return float64(tempo)
 }
